# perf: faster plain string and known value checks