--- a/src/index.ts
+++ b/src/index.ts
@@ -37,7 +37,8 @@ export function destr<T = unknown>(value: any, options: Options = {}): T {
     // eslint-disable-next-line unicorn/prefer-at
     value[0] === '"' &&
     value.endsWith('"') &&
-    !value.includes("\\")
+    // eslint-disable-next-line unicorn/prefer-includes
+    value.indexOf("\\") === -1
   ) {
     return value.slice(1, -1) as T;
   }