I found some minor optimizations to improve performance (mainly with `Bun`).

These are the affected benchmarks taken after each commit compared to `main` (other benchmarks were within margin of error):
- #### perf: move `trim()` after plain string check
```sc
// Node v23.9.0
• plain string (short)
----------------------------------------------------
destr @ dev    11.54 ns/iter      (9.35 ns … 106 ns)
destr @ main   14.82 ns/iter   (12.82 ns … 96.96 ns)

summary for plain string (short)
  destr @ dev
   1.29x faster than destr @ main
   
// Bun v1.2.4
• plain string (short)
----------------------------------------------------
destr @ dev    33.53 ns/iter     (27.88 ns … 735 ns)
destr @ main   39.81 ns/iter     (34.15 ns … 763 ns)

summary for plain string (short)
  destr @ dev
   1.19x faster than destr @ main
```
   
- #### perf: use `indexOf` instead of `includes`
```sc
// Bun v1.2.4
• plain string (short)
----------------------------------------------------
destr @ dev    20.29 ns/iter     (17.53 ns … 709 ns)
destr @ main   38.57 ns/iter     (33.94 ns … 790 ns)

summary for plain string (short)
  destr @ dev
   1.9x faster than destr @ main
```

- #### perf: use `switch` instead of multiple `if` statements
```sc
// Bun v1.2.4
• Known values
----------------------------------------------------
destr @ dev    14.77 ns/iter     (13.94 ns … 648 ns)
destr @ main   23.07 ns/iter     (21.35 ns … 765 ns)

summary for Known values
  destr @ dev
   1.56x faster than destr @ main

• plain string (short)
----------------------------------------------------
destr @ dev     17.7 ns/iter     (15.16 ns … 970 ns)
destr @ main   38.57 ns/iter     (33.66 ns … 769 ns)

summary for plain string (short)
  destr @ dev
   2.18x faster than destr @ main
``` 

- #### perf: replace `endsWith` with `at` for better performance
```sc
// Node v23.9.0
• plain string (short)
----------------------------------------------------
destr @ dev    14.14 ns/iter    (11.88 ns … 93.9 ns)
destr @ main   15.32 ns/iter     (12.83 ns … 116 ns)

summary for plain string (short)
  destr @ dev
   1.08x faster than destr @ main // <----- This is slower than before this commit!
   
// Bun v1.2.4
• plain string (short)
----------------------------------------------------
destr @ dev     11.9 ns/iter     (9.86 ns … 1000 ns)
destr @ main      42 ns/iter     (36.62 ns … 675 ns)

summary for plain string (short)
  destr @ dev
   3.53x faster than destr @ main
```
This is slower in `node v23.9.0` than `endsWith()` to my surprise.
I suspect that this is a bug in `node`, so I also tested with `node v20.10.0` and it's `1.53x faster than destr @ main` there.


- #### perf: improve performance with `node`
```sc
// Node v23.9.0
• plain string (short)
----------------------------------------------------
destr @ main   15.08 ns/iter     (12.75 ns … 104 ns)
destr @ dev    11.45 ns/iter      (9.65 ns … 110 ns)

summary for plain string (short)
  destr @ dev
   1.32x faster than destr @ main
```
`value[value.length - 1]` works as fast as `value.at(-1)` in previous versions of `node` so I changed it to use that.
It's now as fast as `Bun` in that bench.